+++ packages/core/src/shared/identity.ts
@@ -1,4 +1,4 @@
-import { string, z } from "zod/v4"
+import { email, string, z } from "zod/v4"
 import type { EditableShape } from "@/@types/utility.ts"
 
 export type { InferAuthIdentity, InferIdentity, InferShape, ShapeToObject, EditableShape } from "@/@types/utility.ts"
@@ -7,7 +7,7 @@ export const UserIdentity = z.object({
     sub: string(),
     name: string().nullable().optional(),
     image: string().nullable().optional(),
-    email: string().nullable().optional(),
+    email: email().nullable().optional(),
 })
 
 export type UserShape = (typeof UserIdentity)["shape"]